// src/lib/client/model/postShortcutReq.ts
/**
 * Generated by orval v6.21.0 🍺
 * Do not edit manually.
 * Nobstacle API
 * OpenAPI spec version: 1.0.0
 */
import type { PostShortcutReqType } from './postShortcutReqType';

export interface PostShortcutReq {
  extraValue?: string | null;
  key?: string | null;
  order?: number;
  type: PostShortcutReqType;
  value: string;
  color?:string
}